Update from task d2a6e0d8-29d9-4442-b150-09462e70c50e

+++ src/main/java/parsmodernpvp_knl2s7pw/client/ui/PARSSlider.java
@@ -55,7 +55,7 @@ public final class PARSSlider {
          
          // Thumb body
          int thumbColor = -1;
-         fillRoundedRect(g, thumbX, thumbY, thumbSize, thumbSize, thumbSize / 2.0F, thumbColor);
+         fillRoundedRect(g, thumbX, thumbY, thumbSize, thumbSize, (int)(thumbSize / 2.0F), thumbColor);
       }
    }
